Parannettu dokumentaatiota ja jaettu koodia pienempiin palasiin
