+++ src/common.ts
@@ -1,1 +1,3 @@
 export const EXTERNAL_LINK_REGEX = /^https?:\/\//;
+
+export const EMPTY_OBJECT = {} as const;
