+++ src/components/OphInput.tsx
@@ -14,7 +14,8 @@ export type OphInputProps = Omit<
 >;
 
 /**
- * An Input component based on [MUI OutlinedInput](https://mui.com/material-ui/api/outlined-input/).
+ * A text input component based on [MUI OutlinedInput](https://mui.com/material-ui/api/outlined-input/).
+ * If you need label, helper text etc. use [OphInputFormField](/docs/components-ophinputformfield--docs) instead.
  */
 export const OphInput = forwardRef<HTMLInputElement, OphInputProps>(
   function renderInput(props, ref) {
