+++ src/next/OphNextJsLinkBehavior.tsx
@@ -1,9 +1,16 @@
 import { isString } from '@/src/util';
 import NextLink, { type LinkProps } from 'next/link';
 import React from 'react';
-import { EXTERNAL_LINK_REGEX } from '../common';
+import { EXTERNAL_LINK_REGEX } from '@/src/common';
 
-export const LinkBehavior = React.forwardRef<
+/**
+ * A custom link behavior component for Next.js that handles both internal and external links.
+ *
+ * This component uses Next.js Link-component for internal links and a standard a-tag for external links.
+ * External links are identified using a regular expression and are opened in a new tab with
+ * `noopener noreferrer` by default for security.
+ */
+export const OphNextJsLinkBehavior = React.forwardRef<
   HTMLAnchorElement,
   React.PropsWithChildren<LinkProps>
 >(function renderLinkBehavior({ href, children, ...props }, ref) {
