+++ src/next/theme/index.ts
@@ -1,6 +1,1 @@
-export {
-  MUI_NEXTJS_OVERRIDES,
-  openSans,
-  LinkBehavior,
-  OphNextJsThemeProvider,
-} from './theme-nextjs';
+export { OphNextJsThemeProvider } from './OphNextJsThemeProvider';
