+++ src/stories/colors.mdx
@@ -1,7 +1,17 @@
-import { ophColors } from './ophColors';
+import { ophColors } from '../ophColors';
 import { Meta, ColorPalette, ColorItem } from '@storybook/blocks';
 
 <Meta title="Colors" />
+
+# Colors
+
+## Example
+
+```js
+import { ophColors } from '@opetushallitus/oph-design-system';
+
+const BORDER_COLOR = ophColors.grey600;
+```
 
 <ColorPalette>
   <ColorItem
