+++ src/stories/getting-started.mdx
@@ -1,0 +1,69 @@
+import { Meta } from '@storybook/blocks';
+
+<Meta title="Getting started" />
+
+# Getting started
+
+## Personal access token
+
+In order to start using ODS in your project, you first need to create [personal access token for Github Packages](https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token) on github.
+
+## .npmrc file
+
+You need to add the personal access token to your .npmrc file. It makes sense to add it to .npmrc in your home directory.
+Add the following line to the .npmrc file in your home directory:
+
+```
+//npm.pkg.github.com/:_authToken=YOUR_PERSONAL_ACCESS_TOKEN
+```
+
+Replace YOUR_PERSONAL_ACCESS_TOKEN with the token that you generated before.
+
+Create .npmrc file to you project if it doesn't exist yet. Then add following lines to .npmrc file.
+
+```
+@opetushallitus:registry=https://npm.pkg.github.com
+```
+
+## Installation
+
+After configuring the .npmrc file you can install the package with:
+
+`npm install @opetushallitus/oph-design-system`
+
+Also make sure that you have the required peer dependencies installed:
+
+```json
+{
+  "peerDependencies": {
+    "@mui/material": "^6",
+    "@mui/icons-material": "^6",
+    "next": "^14 || ^15", // Mandatory only if you use Next.js
+    "react": "^18 || ^19"
+  }
+}
+```
+
+## Usage
+
+To use ODS components in your project, you need to wrap your App with a Theme provider.
+There are two theme providers available:
+
+1. [OphThemeProvider](/docs/theme-ophthemeprovider--docs) to be used in plain React apps
+2. [OphNextJsThemeProvider](/docs/theme-ophnextjsthemeprovider--docs) to be used in Next.js apps
+
+### Example
+
+`App.tsx`
+
+```jsx
+import React from 'react';
+import { OphThemeProvider } from '@opetushallitus/oph-design-system/theme';
+export const App = () => {
+  return (
+    <OphThemeProvider lang="fi" variant="oph">
+      <Button variant="contained">Button</Button>
+    </Theme>
+  );
+};
+```
