+++ src/theme/theme-utils.ts
@@ -1,7 +1,7 @@
 import { enUS, fiFI, svSE } from '@mui/material/locale';
 import type { OphLanguage } from '../types';
 import { ophColors } from '../ophColors';
-import type { CSSProperties } from 'react';
+import { type CSSProperties } from 'react';
 
 export function getLocale(lang?: OphLanguage) {
   switch (lang) {
